requirements.txt
Django==1.9.2
Markdown==2.6.5
django-filter==0.11.0
djangorestframework==3.3.2
mysql-python==1.2.5
djangorestframework-gis==0.10.1